remind/pyproject.toml
[tool.poetry]
name = "remind"
version = "0.1.0"
description = ""
authors = ["Yuri Kozhanov <ykozhanov97@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
celery = {extras = ["redis"], version = "5.4.0"}
sqlalchemy = "2.0.36"
pydantic = "2.9.2"
requests = "2.32.3"
pytelegrambotapi = "4.23.0"
python-dotenv = "1.0.1"
psycopg2 = "2.9.10"
flower = "2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
